requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
boto3==1.26.144
botocore>=1.29.139
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==14.0
cron-descriptor==1.4.0
cryptography==41.0.1
Django==4.2.1
django-celery-beat==2.5.0
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-import-export==3.2.0
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
dnspython==2.3.0
drf-nested-routers==0.93.4  # https://github.com/alanjds/drf-nested-routers/
drf-spectacular==0.26.2
editdistance==0.6.2
elastic-transport==8.4.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
et-xmlfile==1.1.0
eventlet==0.33.3
factory-boy==3.2.1
Faker==18.9.0
Flask==2.1.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-SocketIO==4.3.2
flask-talisman==1.0.0
g2p==1.1.20230511 # https://github.com/roedoejet/g2p
greenlet==2.0.2
gunicorn==20.1.0
humanfriendly==10.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
munkres==1.1.4
networkx==2.8.4
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
panphon==0.20.0
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-celery==0.0.0
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2023.3
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
requests==2.31.0
rules==3.3
sentry-sdk==1.24.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
text-unidecode==1.3
tqdm==4.65.0
tzdata==2023.3
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.0.3
